feat: make user global state #44
