+++ web/.eslintrc.json
@@ -15,6 +15,7 @@
   "plugins": ["react", "@typescript-eslint"],
   "rules": {
     "no-console": "off",
+    "react/jsx-props-no-spreading": "off",
     "react/jsx-filename-extension": [
       2,
       {
@@ -36,7 +37,8 @@
         "ts": "never",
         "tsx": "never"
       }
-   ]
+   ],
+   "max-len": "off"
   },
   "settings": {
     "import/resolver": {
