+++ web/package.json
@@ -25,6 +25,7 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "6",
     "react-scripts": "^5.0.0",
+    "recoil": "^0.5.2",
     "typescript": "^4.4.2",
     "web-vitals": "^2.1.0"
   },
