+++ web/src/contexts/Contexts.tsx
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const userState = atom({
+  key: 'userState',
+  default: {},
+});
+
+export default userState;
