+++ web/src/templates/SignupTemplate.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import SignupForm from '../organisms/SignupForm';
+
+function SignupTemplate(props: any) {
+  return (
+    <SignupForm
+      {...props}
+    />
+  );
+}
+export default SignupTemplate;
